+++ src/features/product-list/config/index.ts
@@ -1,0 +1,1 @@
+export type { ProductParams } from './type';
