+++ src/features/product-list/config/type.ts
@@ -1,0 +1,6 @@
+export interface ProductParams {
+  pageNumber: number;
+  pageSize: number;
+  sortType?: string;
+  category?: string;
+}
