+++ src/features/user/api/getMyHistoryAuction.ts
@@ -1,6 +1,7 @@
-import { IUserAuctionHistoryList } from "@/@types/AuctionList";
-import { GetAuctionProps } from "@/@types/user";
-import { API_END_POINT, httpClient } from "@/shared";
+import { API_END_POINT, httpClient } from '@/shared';
+
+import { GetAuctionProps } from '@/@types/user';
+import type { IUserAuctionHistoryList } from '@/entities';
 
 export const getMyHistoryAuction = async ({ pageNumber, pageSize }: GetAuctionProps): Promise<IUserAuctionHistoryList> => {
   const response = await httpClient.get(`${API_END_POINT.BID}?status=proceeding&page=${pageNumber}&size=${pageSize}`);
